Refactor: elimilate duplication
